+++ src/Bem/Utl/Utl.hs
@@ -1,0 +1,12 @@
+{-# LANGUAGE EmptyDataDeriving #-}
+
+
+-- | high-level extra utilities
+module Bem.Utl.Utl where
+
+
+-- | Denote absent elements.
+data NoElem m
+
+-- | Denote absent modifiers.
+data NoMod deriving Show