+++ src/Bem/Utl/Utl.hs
@@ -1,8 +1,13 @@
 {-# LANGUAGE EmptyDataDeriving #-}
+{-# LANGUAGE RankNTypes #-}
 
 
 -- | high-level extra utilities
 module Bem.Utl.Utl where
+
+
+import Bem.Cls.Gen.Gen
+import Bem.Utl.Intr
 
 
 -- | Denote absent elements.
@@ -10,3 +15,24 @@ data NoElem m
 
 -- | Denote absent modifiers.
 data NoMod deriving Show
+
+
+-- | Generate a class of a block without its modifiers and full element.
+genNoBlkModsBlk :: FromBlk b (FromFullElem b Class)
+genNoBlkModsBlk blk prntBlk elem' elemMods
+    =
+    genBlkElem blk [] prntBlk elem' elemMods
+
+-- | Generate a class of a block and element that is without its modifiers.
+genNoElemModsBlk :: FromFullBlk b (FromElem b Class)
+genNoElemModsBlk blk blkMods prntBlk elem'
+    =
+    genBlkElem blk blkMods prntBlk elem' []
+
+-- | Generate a class of a block and element without any modifiers.
+genNoModsBlk :: FromBlk b (FromElem b Class)
+genNoModsBlk blk prntBlk elem' = genBlkElem blk [] prntBlk elem' []
+
+-- | Generate a class of an element that is without its modifiers.
+genNoModsElem :: FromElem b Class
+genNoModsElem prntBlk elem' = genElem prntBlk elem' []